+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.4.0
+
+### New features
+
+- ERC20 Claim Window Merkle Claim contract.
+
 ## 3.3.0
 
 ### New features